Add sequential implementation of search supporting sentinels
based on cppreference possible implementation
https://en.cppreference.com/w/cpp/algorithm/ranges/search
